+++ src/library/DIPS.Mobile.UI/API/Camera/ImageCapturing/Android/ImageCapture.cs
@@ -1,3 +1,5 @@
+using Android.Graphics;
+using Android.Hardware.Camera2;
 using Android.Views;
 using AndroidX.Camera.Core;
 using AndroidX.Camera.Core.Internal.Utils;
@@ -48,7 +50,9 @@ public partial class ImageCapture : CameraFragment
         if (Context is null) 
             return;
 
-        m_cameraCaptureUseCase.FlashMode = m_flashActive
+        _ = OnBeforeCapture();
+        
+        m_cameraCaptureUseCase.FlashMode = FlashActive
             ? AndroidX.Camera.Core.ImageCapture.FlashModeOn
             : AndroidX.Camera.Core.ImageCapture.FlashModeOff;
          
@@ -84,44 +88,42 @@ public partial class ImageCapture : CameraFragment
     {
         var imageData = ImageUtil.JpegImageToJpegByteArray(imageProxy);
         var bitmap = imageProxy.ToBitmap();
+        
         using var imageMemoryStream = new MemoryStream(imageData);
         var exif = new ExifInterface(imageMemoryStream);
-        var (orientationConstant, orientationDisplayName) = GetOrientationMetadata(exif);
-        var imageTransformation = new ImageTransformation(orientationConstant, orientationDisplayName);
+        var orientationDegree = exif.ToTrueOrientationDegree();
+        var imageTransformation = new ImageTransformation(orientationDegree, orientationDegree.ToString());
         var thumbnail = await TryGetThumbnail(exif, imageTransformation);
+
+        var tuple = await CapturedImage.RotateBitmapImageBasedOnOrientation(imageTransformation, bitmap);
+
+        imageData = tuple.Item1;
+        bitmap = tuple.Item2;
         
-        var capturedImage = new CapturedImage(imageData,bitmap,  thumbnail, imageProxy.ImageInfo, imageProxy.Width, imageProxy.Height,imageTransformation);
+        var capturedImage = new CapturedImage(imageData, bitmap, thumbnail.Item1, thumbnail.Item2, imageProxy, imageTransformation);
         
-        GoToConfirmState(capturedImage, m_imageCaptureSettings);
+        GoToConfirmState(capturedImage);
     }
 
-    private async Task<byte[]?> TryGetThumbnail(ExifInterface exif, ImageTransformation transformation)
+    private async Task<(byte[], Bitmap)> TryGetThumbnail(ExifInterface exif, ImageTransformation transformation)
     {
-        if (!exif.HasThumbnail) return null;
+        if (!exif.HasThumbnail)
+            return (null!, null!);
 
         var bitmapImage = exif.ThumbnailBitmap;
-        if (bitmapImage == null) return null;
+        if (bitmapImage == null) 
+            return (null!, null!);
         
-        return await CapturedImage.RotateBitmapImageBasedOnOrientation(transformation, bitmapImage); ;
+        return (await CapturedImage.RotateBitmapImageBasedOnOrientationAsByteArray(transformation, bitmapImage));
     }
 
-    private static (int orientationConstant, string orientationDisplayName) GetOrientationMetadata(ExifInterface exif)
+    private static int GetOrientationMetadata(ExifInterface exif)
     {
+        exif.SetAttribute(ExifInterface.TagOrientation, string.Empty);
+
         var orientationConstant = exif.GetAttributeInt(ExifInterface.TagOrientation, ExifInterface.OrientationNormal);
         
-        var orientationDisplayName = orientationConstant switch
-        {
-            ExifInterface.OrientationNormal => nameof(ExifInterface.OrientationNormal),
-            ExifInterface.OrientationRotate90 => nameof(ExifInterface.OrientationRotate90),
-            ExifInterface.OrientationRotate180 => nameof(ExifInterface.OrientationRotate180),
-            ExifInterface.OrientationRotate270 => nameof(ExifInterface.OrientationRotate270),
-            ExifInterface.OrientationTranspose => nameof(ExifInterface.OrientationTranspose),
-            ExifInterface.OrientationTransverse => nameof(ExifInterface.OrientationTransverse),
-            ExifInterface.OrientationUndefined => nameof(ExifInterface.OrientationUndefined),
-            _ => throw new ArgumentOutOfRangeException(nameof(orientationConstant))
-        };
-
-        return (orientationConstant, orientationDisplayName);
+        return orientationConstant;
     }
 
     private static long[] GetBitsPerSample(ExifInterface exif)
@@ -152,6 +154,8 @@ internal class ImageCaptureCallback(
         base.OnCaptureSuccess(image);
         image?.Close();
     }
+    
+    
 
     protected override void Dispose(bool disposing)
     {