+++ src/library/DIPS.Mobile.UI/API/Camera/ImageCapturing/Observers/IEditStateObserver.cs
@@ -1,0 +1,8 @@
+namespace DIPS.Mobile.UI.API.Camera.ImageCapturing.Observers;
+
+internal interface IImageEditStateObserver
+{
+    void OnSaveButtonTapped();
+    void OnCancelButtonTapped();
+    Task OnRotateButtonTapped(bool clockwise);
+}