+++ src/library/DIPS.Mobile.UI/API/Camera/ImageCapturing/Views/BottomToolbar/ConfirmState/ConfirmStateView.cs
@@ -1,73 +1,30 @@
+using DIPS.Mobile.UI.API.Camera.ImageCapturing.Observers;
 using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
-using DIPS.Mobile.UI.Resources.Styles;
-using DIPS.Mobile.UI.Resources.Styles.Button;
-using DIPS.Mobile.UI.Resources.Styles.Label;
-using Button = DIPS.Mobile.UI.Components.Buttons.Button;
-using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
-using VerticalStackLayout = DIPS.Mobile.UI.Components.Lists.VerticalStackLayout;
 
 namespace DIPS.Mobile.UI.API.Camera.ImageCapturing.Views.BottomToolbar.ConfirmState;
 
-public class ConfirmStateView : Grid
+internal class ConfirmStateView : Grid
 {
-    public ConfirmStateView(Action? usePhoto, Action? retakePhoto)
+    public ConfirmStateView(IConfirmStateObserver confirmStateObserver)
     {
         AddColumnDefinition(new ColumnDefinition(GridLength.Star));
         AddColumnDefinition(new ColumnDefinition(GridLength.Star));
         
         VerticalOptions = LayoutOptions.Center;
-        
-        var retakeButton = new Button
-        {
-            ImageSource = Icons.GetIcon(IconName.arrow_back_line),
-            ImageTintColor = Colors.GetColor(ColorName.color_system_white),
-            BackgroundColor = Colors.GetColor(ColorName.color_neutral_90),
-            Style = Styles.GetButtonStyle(ButtonStyle.GhostIconButtonLarge),
-            Command = new Command(retakePhoto)
-        };
-        
-        var doneButton = new Button
-        {
-            ImageSource = Icons.GetIcon(IconName.check_line),
-            ImageTintColor = Colors.GetColor(ColorName.color_system_white),
-            BackgroundColor = Colors.GetColor(ColorName.color_neutral_90),
-            Style = Styles.GetButtonStyle(ButtonStyle.GhostIconButtonLarge),
-            Command = new Command(usePhoto)
-        };
-        
-        var doneLabel = new Label
-        {
-            Text = DUILocalizedStrings.UsePicture,
-            Style = Styles.GetLabelStyle(LabelStyle.UI100),
-            TextColor = Colors.GetColor(ColorName.color_system_white),
-            HorizontalOptions = LayoutOptions.Center,
-            Margin = new Thickness(0, Sizes.GetSize(SizeName.size_3), 0, 0)
-        };
-        
-        var retakeLabel = new Label
-        {
-            Text = DUILocalizedStrings.ReTakePhoto,
-            Style = Styles.GetLabelStyle(LabelStyle.UI100),
-            TextColor = Colors.GetColor(ColorName.color_system_white),
-            HorizontalOptions = LayoutOptions.Center,
-            Margin = new Thickness(0, Sizes.GetSize(SizeName.size_3), 0, 0)
-        };
 
-        var leftColumn = new VerticalStackLayout
-        {
-            Spacing = 0,
-            Children = { retakeButton, retakeLabel },
-            HorizontalOptions = LayoutOptions.Start
-        };
+        var doneButton =
+            new ButtonWithText(DUILocalizedStrings.UsePicture, Icons.GetIcon(IconName.check_line), confirmStateObserver.OnUsePhotoButtonTapped)
+            {
+                HorizontalOptions = LayoutOptions.End
+            };
+
+        var retakeButton =
+            new ButtonWithText(DUILocalizedStrings.ReTakePhoto, Icons.GetIcon(IconName.arrow_back_line), confirmStateObserver.OnRetakePhotoButtonTapped)
+            {
+                HorizontalOptions = LayoutOptions.Start
+            };
         
-        var rightColumn = new VerticalStackLayout
-        {
-            Spacing = 0,
-            Children = { doneButton, doneLabel },
-            HorizontalOptions = LayoutOptions.End
-        };
-        
-        Add(leftColumn);
-        this.Add(rightColumn, 1);
+        Add(retakeButton);
+        this.Add(doneButton, 1);
     }
 }